requirements.txt
# Core FastAPI
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# AI/ML
huggingface-hub==0.20.2

# Image Processing
Pillow==10.2.0

# Utilities
python-jose[cryptography]==3.3.0  # For JWT token validation
